+++ tests/FizzWare.NBuilder.Tests/Integration/ListBuilderTests.cs
@@ -12,16 +12,11 @@ using Xunit;
 
 namespace FizzWare.NBuilder.Tests.Integration
 {
-    /// <remarks>
-    /// To run these tests, create a local database named 'NBuilderTests'
-    /// </remarks>
-
     public class ListBuilderTests
     {
         public ListBuilderTests()
         {
             new RepositoryBuilderSetup().DoSetup();
-
         }
 
         [Theory]
@@ -30,8 +25,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [InlineData(9)]
         public void SetValueOfPropertyForGivenIndexObject(int index)
         {
-            var builderSetup = new BuilderSettings();
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                            .CreateListOfSize<Product>(10)
                            .IndexOf(index)
                            .With(x => x.Title = "A special title")
@@ -50,8 +44,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         public void SetValueOfPropertyForGivenIndexesObject()
         {
             int[] indexes = new[] { 0, 2, 9 };
-            var builderSetup = new BuilderSettings();
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                            .CreateListOfSize<Product>(10)
                            .IndexOf(indexes)
                            .With(x => x.Title = "A special title")
@@ -72,8 +65,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void CreatingAList()
         {
-            var builderSetup = new BuilderSettings();
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                 .CreateListOfSize<Product>(10)
                 .Build()
                 ;
@@ -106,8 +98,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingAllToSetValues()
         {
-            var builderSetup = new BuilderSettings();
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                 .CreateListOfSize<Product>(10)
                 .All()
                     .With(x => x.Title = "A special title")
@@ -120,8 +111,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void SettingTheValueOfAProperty()
         {
-            var builderSetup = new BuilderSettings();
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                 .CreateListOfSize<Product>(10)
                 .TheFirst(2)
                     .With(x => x.Title = "A special title")
@@ -140,8 +130,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingSingularSyntaxInstead()
         {
-            var builderSetup = new BuilderSettings();
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                             .CreateListOfSize<Product>(10)
                             .TheFirst(1)
                                 .With(x => x.Title = "Special title 1")
@@ -153,10 +142,9 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingAGenerator()
         {
-            var builderSetup = new BuilderSettings();
             var generator = new RandomGenerator();
 
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                 .CreateListOfSize<Product>(10)
                 .All()
                     .With(x => x.PriceBeforeTax = generator.Next(50, 1000))
@@ -187,10 +175,10 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void DeclaringThatARandomNumberOfElementsShouldHaveCertainProperties()
         {
-            var builderSetup = new BuilderSettings();
             const string specialdescription = "SpecialDescription";
             const decimal specialPrice = 10m;
-            var products = new Builder(builderSetup).CreateListOfSize<Product>(10)
+            var products = new Builder()
+                .CreateListOfSize<Product>(10)
                 .Random(5)
                     .With(x => x.Description = specialdescription)
                     .And(x => x.PriceBeforeTax = specialPrice)
@@ -213,9 +201,10 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void CreatingAListOfProductsAndAddingThemToCategories()
         {
-            var builderSetup = new BuilderSettings();
-            var builder = new Builder(builderSetup);
-            var categories = builder.CreateListOfSize<Category>(50).Build();
+            var builder = new Builder();
+            var categories = builder
+                .CreateListOfSize<Category>(50)
+                .Build();
 
             var products = builder
                             .CreateListOfSize<Product>(500)
@@ -233,11 +222,14 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingTheSequentialGenerator()
         {
-            var builderSetup = new BuilderSettings();
-            var generator = new SequentialGenerator<int> { Direction = GeneratorDirection.Descending, Increment = 2 };
+            var generator = new SequentialGenerator<int>
+            {
+                Direction = GeneratorDirection.Descending,
+                Increment = 2
+            };
             generator.StartingWith(6);
 
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                 .CreateListOfSize<Product>(3)
                 .All()
                     .With(x => x.Id = generator.Generate())
@@ -252,7 +244,6 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingSequentialGenerators()
         {
-            var builderSetup = new BuilderSettings();
             // Arrange
             var decimalGenerator = new SequentialGenerator<decimal>
             {
@@ -262,10 +253,14 @@ namespace FizzWare.NBuilder.Tests.Integration
 
             decimalGenerator.StartingWith(2000);
 
-            var intGenerator = new SequentialGenerator<int> { Increment = 10000 };
+            var intGenerator = new SequentialGenerator<int>
+            {
+                Increment = 10000
+            };
 
             // Act
-            var list = new Builder(builderSetup).CreateListOfSize<Product>(3)
+            var list = new Builder()
+                .CreateListOfSize<Product>(3)
                 .All()
                     .With(x => x.PriceBeforeTax = decimalGenerator.Generate())
                     .And(x => x.Id = intGenerator.Generate())
@@ -284,7 +279,6 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void SequentialGenerator_DateTimeGeneration()
         {
-            var builderSetup = new BuilderSettings();
             const int increment = 2;
             var dateTimeGenerator = new SequentialGenerator<DateTime>
             {
@@ -297,7 +291,8 @@ namespace FizzWare.NBuilder.Tests.Integration
             dateTimeGenerator.StartingWith(startingDate);
 
 
-            var list = new Builder(builderSetup).CreateListOfSize<Product>(2)
+            var list = new Builder()
+                .CreateListOfSize<Product>(2)
                 .All()
                     .With(x => x.Created = dateTimeGenerator.Generate())
                 .Build();
@@ -309,9 +304,10 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingTheWithBetween_And_SyntaxForGreaterReadability()
         {
-            var builderSetup = new BuilderSettings();
-            var builder = new Builder(builderSetup);
-            var categories = builder.CreateListOfSize<Category>(50).Build();
+            var builder = new Builder();
+            var categories = builder
+                .CreateListOfSize<Category>(50)
+                .Build();
 
             var products = builder
                             .CreateListOfSize<Product>(500)
@@ -329,10 +325,9 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void CreatingAListOfATypeWithAConstructor()
         {
-            var builderSetup = new BuilderSettings();
             // ctor: BasketItem(ShoppingBasket basket, Product product, int quantity)
 
-            var builder = new Builder(builderSetup);
+            var builder = new Builder();
             var basket = builder.CreateNew<ShoppingBasket>().Build();
             var product = builder.CreateNew<Product>().Build();
             const int quantity = 5;
@@ -354,7 +349,6 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void DifferentPartsOfTheListCanBeConstructedDifferently()
         {
-            var builderSetup = new BuilderSettings();
             var basket1 = new ShoppingBasket();
             var product1 = new Product();
             const int quantity1 = 5;
@@ -363,7 +357,7 @@ namespace FizzWare.NBuilder.Tests.Integration
             var product2 = new Product();
             const int quantity2 = 7;
 
-            var items = new Builder(builderSetup)
+            var items = new Builder()
                 .CreateListOfSize<BasketItem>(4)
                 .TheFirst(2)
                     .WithConstructor(() => new BasketItem(basket1, product1, quantity1))
@@ -384,16 +378,16 @@ namespace FizzWare.NBuilder.Tests.Integration
         {
             Should.Throw<BuilderException>(() =>
             {
-                var builderSetup = new BuilderSettings();
-                new Builder(builderSetup).CreateListOfSize<ChuckNorris>(10).Build();
+                new Builder()
+                    .CreateListOfSize<ChuckNorris>(10)
+                    .Build();
             });
         }
 
         [Fact]
         public void UsingWith_WithImmutableClassProperties()
         {
-            var builderSetup = new BuilderSettings();
-            var invoices = new Builder(builderSetup)
+            var invoices = new Builder()
                 .CreateListOfSize<Invoice>(2)
                 .TheFirst(2)
                     .With(p => p.Amount, 100)
@@ -406,8 +400,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingWith_WithAnIndex()
         {
-            var builderSetup = new BuilderSettings();
-            var invoices = new Builder(builderSetup)
+            var invoices = new Builder()
                 .CreateListOfSize<Product>(2)
                 .TheFirst(2)
                     .With((p, idx) => p.Description = "Description" + (idx + 5))
@@ -420,8 +413,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingAndWithAnIndex()
         {
-            var builderSetup = new BuilderSettings();
-            var invoices = new Builder(builderSetup)
+            var invoices = new Builder()
                 .CreateListOfSize<Product>(2)
                 .TheFirst(2)
                     .With(p => p.Title = "Title")
@@ -437,8 +429,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingAndWithImmutableClassProperties()
         {
-            var builderSetup = new BuilderSettings();
-            var invoices = new Builder(builderSetup)
+            var invoices = new Builder()
                 .CreateListOfSize<Invoice>(2)
                 .TheFirst(2)
                     .With(p => p.Amount, 100)
@@ -454,8 +445,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingHasWithImmutableClassProperties()
         {
-            var builderSetup = new BuilderSettings();
-            var invoices = new Builder(builderSetup)
+            var invoices = new Builder()
                 .CreateListOfSize<Invoice>(1)
                 .TheFirst(1)
                     .With(p => p.Amount, 100)
@@ -468,8 +458,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void ChainingDeclarationsTogether()
         {
-            var builderSetup = new BuilderSettings();
-            var list = new Builder(builderSetup)
+            var list = new Builder()
                 .CreateListOfSize<Product>(30)
                 .TheFirst(10)
                     .With(x => x.Title = "Special Title 1")
@@ -487,28 +476,10 @@ namespace FizzWare.NBuilder.Tests.Integration
             list[29].Title.ShouldBe("Special Title 3");
         }
 
-        ////[Fact]
-        ////public void UsingAndTheRemaining()
-        ////{ 
-        ////    var list = Builder<Product>
-        ////        .CreateListOfSize(4)
-        ////        .TheFirst(2)
-        ////            .With(x => x.Title = "Special Title 1")
-        ////        .TheRemainder()
-        ////            .With(x => x.Title = "Special Title 2")
-        ////        .Build();
-
-        ////    list[0].Title.ShouldBe("Special Title 1");
-        ////    list[1].Title.ShouldBe("Special Title 1");
-        ////    list[2].Title.ShouldBe("Special Title 2");
-        ////    list[3].Title.ShouldBe("Special Title 2");
-        ////}
-
         [Fact]
         public void UsingAndThePrevious()
         {
-            var builderSetup = new BuilderSettings();
-            var list = new Builder(builderSetup)
+            var list = new Builder()
                 .CreateListOfSize<Product>(30)
                 .TheLast(10)
                     .With(x => x.Title = "Special Title 1")
@@ -525,8 +496,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingSection()
         {
-            var builderSetup = new BuilderSettings();
-            var list = new Builder(builderSetup)
+            var list = new Builder()
                 .CreateListOfSize<Product>(30)
                 .All()
                     .With(x => x.Title = "Special Title 1")
@@ -556,8 +526,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingSectionAndTheNext()
         {
-            var builderSetup = new BuilderSettings();
-            var list = new Builder(builderSetup)
+            var list = new Builder()
                 .CreateListOfSize<Product>(30)
                 .All()
                     .With(x => x.Title = "Special Title 1")
@@ -576,7 +545,9 @@ namespace FizzWare.NBuilder.Tests.Integration
         public void UsingDo()
         {
             var builder = new Builder();
-            var children = builder.CreateListOfSize<Category>(3).Build();
+            var children = builder
+                .CreateListOfSize<Category>(3)
+                .Build();
 
             var categories = builder
                 .CreateListOfSize<Category>(10)
@@ -598,10 +569,11 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingDoAndPickTogether()
         {
-            var builderSetup = new BuilderSettings();
-            var children = new Builder(builderSetup).CreateListOfSize<Category>(10).Build();
+            var children = new Builder()
+                .CreateListOfSize<Category>(10)
+                .Build();
 
-            var categories = new Builder(builderSetup)
+            var categories = new Builder()
                 .CreateListOfSize<Category>(10)
                 .TheFirst(2)
                     .Do(x => x.AddChild(Pick<Category>.RandomItemFrom(children)))
@@ -615,14 +587,16 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Description("Multi functions allow you to call a method on an object on each item in a list")]
         public void UsingMultiFunctions()
         {
-            var builderSetup = new BuilderSettings();
-            var categories = new Builder(builderSetup).CreateListOfSize<Category>(5).Build();
+            var categories = new Builder()
+                .CreateListOfSize<Category>(5)
+                .Build();
 
             // Here we are saying, add every product to all of the categories
 
-            var products = new Builder(builderSetup)
+            var products = new Builder()
                 .CreateListOfSize<Product>(10)
-                .All().DoForEach((product, category) => product.AddToCategory(category), categories)
+                .All()
+                .DoForEach((product, category) => product.AddToCategory(category), categories)
                 .Build();
 
             // Assertions are intentionally verbose for clarity
@@ -640,10 +614,10 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void UsingRandomGenerator()
         {
-            var builderSetup = new BuilderSettings();
             var generator = new RandomGenerator();
 
-            var list = new Builder(builderSetup).CreateListOfSize<Product>(3)
+            var list = new Builder()
+                .CreateListOfSize<Product>(3)
                 .All()
                 .With(x => x.QuantityInStock = generator.Next(1000, 2000))
                 .Build();
@@ -661,11 +635,10 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void WillNotLetYouDoThingsThatDoNotMakeSense()
         {
-            var builderSetup = new BuilderSettings();
 
             Should.Throw<BuilderException>(() =>
             {
-                new Builder(builderSetup)
+                new Builder()
                     .CreateListOfSize<Product>(10)
                     .TheFirst(5)
                         .With(x => x.Title = "titleone")
@@ -679,8 +652,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         [Fact]
         public void SupportsStructsButDoesNotSupportAutomaticallyNamingTheProperties()
         {
-            var builderSetup = new BuilderSettings();
-            var locations = new Builder(builderSetup)
+            var locations = new Builder()
                 .CreateListOfSize<WarehouseLocation>(10)
                 .Build();
 
@@ -716,7 +688,9 @@ namespace FizzWare.NBuilder.Tests.Integration
             {
                 AutoNameProperties = false
             };
-            var products = new Builder(builderSetup).CreateListOfSize<Product>(10).Build();
+            var products = new Builder(builderSetup)
+                .CreateListOfSize<Product>(10)
+                .Build();
 
             products[0].Title.ShouldBeNull();
             products[9].Title.ShouldBeNull();
@@ -747,7 +721,7 @@ namespace FizzWare.NBuilder.Tests.Integration
         {
 
             var builderSetup = new BuilderSettings();
-            var propertyNamer = new RandomValuePropertyNamer(new RandomGenerator(), new ReflectionUtil(), true, DateTime.Now, DateTime.Now.AddDays(10), true, new BuilderSettings());
+            var propertyNamer = new RandomValuePropertyNamer(builderSetup);
             builderSetup.SetDefaultPropertyNamer(propertyNamer);
 
             var products = new Builder(builderSetup).CreateListOfSize<Product>(10).Build();